fixed some linting issues
